Students management, result creation and report card for selection finalized

+++ src/composables/useDisplayDate.ts
@@ -4,7 +4,7 @@ export const useDisplayDate = () => {
   const formattedDate = ref('')
 
   function displayDate(dateString) {
-    const options = { year: 'numeric', month: 'long', day: 'numeric' }
+    const options = { year: 'numeric', month: 'numeric', day: 'numeric' }
     const date = new Date(dateString)
     formattedDate.value = date.toLocaleDateString('en-IN', options)
   }
